format, seperate tags
